Moved de haser function into its own crate to be inserted into a lambda
