+++ backend/hash_slinging_slasher/de_hasher/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "de_hasher"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde_json = "1.0.107"
+serde = { version = "1.0.188", features = ["derive"] }
+flate2 = "1.0.28"
+bincode="1.3.3"
