+++ backend/hash_slinging_slasher/hasher/src/main.rs
@@ -1,5 +1,5 @@
 use std::error::Error;
-use std::time::{SystemTime, UNIX_EPOCH, Instant, Duration};
+use std::time::Duration;
 use serde::{Deserialize, Serialize};
 use std::fmt::Formatter;
 use std::fmt::Display;
@@ -7,8 +7,7 @@ use std::fs::File;
 use std::io::Read;
 use std::io::Write;
 use flate2::Compression;
-use flate2::write::ZlibEncoder;
-use flate2::read::ZlibDecoder;
+
 
 
 
@@ -46,21 +45,10 @@ fn main() {
     let data = retrieve_data().expect("Failed to get data");
     let c_data = encrypt_data(data).expect("Failed to compress data");
     println!("{:?}", &c_data);
-    decrypt_data(c_data).unwrap();
 }
 
 
-fn decrypt_data(compressed_data: Vec<u8>) -> Result<EventLogData, Box<dyn Error>> {
-    let mut decompressed_data = Vec::new();
-    let mut decoder = ZlibDecoder::new(&compressed_data[..]);
-    decoder.read_to_end(&mut decompressed_data).expect("Decompression failed");
 
-    // Step 4: Deserialize the data from binary format
-    let deserialized_data: EventLogData = bincode::deserialize(&decompressed_data).expect("Deserialization failed");
-
-    // 'deserialized_data' now contains the original data
-    Ok(deserialized_data)
-}
 
 fn encrypt_data(event_data: EventLogData) -> Result<Vec<u8>, Box<dyn Error>> {
     let serialized_data = bincode::serialize(&event_data).expect("Serialization failed");
@@ -101,12 +89,6 @@ fn json_retrieve() -> Result<EventLogData, Box<dyn Error>> {
 mod tests {
 
     use super::*;
-    // #[test]
-    // #[ignore]
-    // fn test_retrieve_data() {
-    //
-    //     assert_eq!(retrieve_data().unwrap(), ());
-    // }
 
     #[test]
     fn test_json_retrieve() {
@@ -150,28 +132,6 @@ mod tests {
         let expected_data = vec![120, 156, 99, 98, 64, 128, 61, 75, 180, 82, 65, 180, 4, 155, 142, 0, 11, 144, 254, 178, 20, 194, 127, 81, 100, 32, 0, 0, 88, 77, 6, 18];
 
         let json_data = encrypt_data(event_data);
-
-        assert_eq!(json_data.unwrap(), expected_data);
-    }
-
-    #[test]
-    fn test_decryption() {
-
-        let expected_data = EventLogData {
-            data: vec![
-                EventLog {
-                    change: ChangeEvent::Up,
-                    time: Duration::new(1697293500, 271320600)
-                },
-                EventLog {
-                    change: ChangeEvent::A,
-                    time: Duration::new(1697293812, 271610600)
-                },
-            ],
-
-        };
-
-        let json_data = decrypt_data(vec![120, 156, 99, 98, 64, 128, 61, 75, 180, 82, 65, 180, 4, 155, 142, 0, 11, 144, 254, 178, 20, 194, 127, 81, 100, 32, 0, 0, 88, 77, 6, 18]);
 
         assert_eq!(json_data.unwrap(), expected_data);
     }
